+++ cms_helper.py
@@ -1,35 +1,31 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
+from __future__ import absolute_import, print_function, unicode_literals
 
-import sys
 from tempfile import mkdtemp
+
 gettext = lambda s: s
 
-HELPER_SETTINGS = {
-    'NOSE_ARGS': [
-        '-s',
-    ],
-    'ROOT_URLCONF': 'tests.test_utils.urls',
-    'INSTALLED_APPS': [
-        'admin_enhancer',
+HELPER_SETTINGS = dict(
+    ROOT_URLCONF='tests.test_utils.urls',
+    INSTALLED_APPS=[
         'filer',
         'parler',
-        'taggit',
         'meta',
         'meta_mixin',
         'easy_thumbnails',
         'djangocms_text_ckeditor',
         'cmsplugin_filer_image',
-        'django_select2',
+        'taggit',
         'taggit_autosuggest',
     ],
-    'LANGUAGE_CODE': 'en',
-    'LANGUAGES': (
+    LANGUAGE_CODE='en',
+    LANGUAGES=(
         ('en', gettext('English')),
         ('fr', gettext('French')),
         ('it', gettext('Italiano')),
     ),
-    'CMS_LANGUAGES': {
+    CMS_LANGUAGES={
         1: [
             {
                 'code': 'en',
@@ -58,7 +54,7 @@ HELPER_SETTINGS = {
             'hide_untranslated': False,
         },
     },
-    'PARLER_LANGUAGES': {
+    PARLER_LANGUAGES={
         1: (
             {'code': 'en'},
             {'code': 'it'},
@@ -72,35 +68,33 @@ HELPER_SETTINGS = {
             'hide_untranslated': False,
         }
     },
-    'MIDDLEWARE_CLASSES': [
-        'django.contrib.messages.middleware.MessageMiddleware',
-    ],
-    'MIGRATION_MODULES': {
-        'filer': 'filer.migrations_django',
+    MIGRATION_MODULES={
         'cmsplugin_filer_image': 'cmsplugin_filer_image.migrations_django',
-        'djangocms_text_ckeditor': 'djangocms_text_ckeditor.migrations_django',
     },
-    'META_SITE_PROTOCOL': 'http',
-    'META_SITE_DOMAIN': 'example.com',
-    'META_USE_OG_PROPERTIES': True,
-    'META_USE_TWITTER_PROPERTIES': True,
-    'META_USE_GOOGLEPLUS_PROPERTIES': True,
-    'THUMBNAIL_PROCESSORS': (
+    META_SITE_PROTOCOL='http',
+    META_SITE_DOMAIN='example.com',
+    META_USE_OG_PROPERTIES=True,
+    META_USE_TWITTER_PROPERTIES=True,
+    META_USE_GOOGLEPLUS_PROPERTIES=True,
+    THUMBNAIL_PROCESSORS=(
         'easy_thumbnails.processors.colorspace',
         'easy_thumbnails.processors.autocrop',
         'filer.thumbnail_processors.scale_and_crop_with_subject_location',
         'easy_thumbnails.processors.filters',
     ),
-    'FILE_UPLOAD_TEMP_DIR': mkdtemp(),
-    'SITE_ID': 1
-}
-if 'test' in sys.argv:
-    HELPER_SETTINGS['INSTALLED_APPS'].append('django_nose')
+    FILE_UPLOAD_TEMP_DIR=mkdtemp(),
+    SITE_ID=1
+)
+
+try:
+    import admin_enhancer  # pragma: no cover # NOQA
+    HELPER_SETTINGS['INSTALLED_APPS'].append('admin_enhancer')
+except ImportError:
+    pass
 
 
 def run():
     from djangocms_helper import runner
-    sys.argv.append('--nose-runner')
     runner.cms('djangocms_blog')
 
 if __name__ == "__main__":