+++ djangocms_blog/cms_plugins.py
@@ -1,12 +1,13 @@
 # -*- coding: utf-8 -*-
-from django.utils.translation import ugettext_lazy as _
+from __future__ import absolute_import, print_function, unicode_literals
 
 from cms.models.pluginmodel import CMSPlugin
 from cms.plugin_base import CMSPluginBase
 from cms.plugin_pool import plugin_pool
+from django.utils.translation import ugettext_lazy as _
 
-from .models import AuthorEntriesPlugin, LatestPostsPlugin, Post, BlogCategory
 from .forms import LatestEntriesForm
+from .models import AuthorEntriesPlugin, BlogCategory, LatestPostsPlugin, Post
 from .settings import get_setting
 
 
@@ -24,10 +25,11 @@ class BlogLatestEntriesPlugin(BlogPlugin):
     model = LatestPostsPlugin
     form = LatestEntriesForm
     filter_horizontal = ('categories',)
+    fields = ('latest_posts', 'tags', 'categories')
     cache = False
 
     def render(self, context, instance, placeholder):
-        context['instance'] = instance
+        context = super(BlogLatestEntriesPlugin, self).render(context, instance, placeholder)
         context['posts_list'] = instance.get_posts(context['request'])
         context['TRUNCWORDS_COUNT'] = get_setting('POSTS_LIST_TRUNCWORDS_COUNT')
         return context
@@ -42,9 +44,10 @@ class BlogLatestEntriesPluginCached(BlogPlugin):
     model = LatestPostsPlugin
     form = LatestEntriesForm
     filter_horizontal = ('categories',)
+    fields = ('latest_posts', 'tags', 'categories')
 
     def render(self, context, instance, placeholder):
-        context['instance'] = instance
+        context = super(BlogLatestEntriesPluginCached, self).render(context, instance, placeholder)
         context['posts_list'] = instance.get_posts()
         context['TRUNCWORDS_COUNT'] = get_setting('POSTS_LIST_TRUNCWORDS_COUNT')
         return context
@@ -59,7 +62,7 @@ class BlogAuthorPostsPlugin(BlogPlugin):
     filter_horizontal = ['authors']
 
     def render(self, context, instance, placeholder):
-        context['instance'] = instance
+        context = super(BlogAuthorPostsPlugin, self).render(context, instance, placeholder)
         context['authors_list'] = instance.get_authors()
         return context
 
@@ -71,6 +74,7 @@ class BlogTagsPlugin(BlogPlugin):
     render_template = 'djangocms_blog/plugins/tags.html'
 
     def render(self, context, instance, placeholder):
+        context = super(BlogTagsPlugin, self).render(context, instance, placeholder)
         context['tags'] = Post.objects.tag_cloud(queryset=Post.objects.published())
         return context
 
@@ -82,6 +86,7 @@ class BlogCategoryPlugin(BlogPlugin):
     render_template = 'djangocms_blog/plugins/categories.html'
 
     def render(self, context, instance, placeholder):
+        context = super(BlogCategoryPlugin, self).render(context, instance, placeholder)
         context['categories'] = BlogCategory.objects.all()
         return context
 
@@ -93,6 +98,7 @@ class BlogArchivePlugin(BlogPlugin):
     render_template = 'djangocms_blog/plugins/archive.html'
 
     def render(self, context, instance, placeholder):
+        context = super(BlogArchivePlugin, self).render(context, instance, placeholder)
         context['dates'] = Post.objects.get_months(queryset=Post.objects.published())
         return context