+++ djangocms_blog/cms_toolbar.py
@@ -1,9 +1,10 @@
 # -*- coding: utf-8 -*-
+from __future__ import absolute_import, print_function, unicode_literals
+
 from cms.toolbar_base import CMSToolbar
 from cms.toolbar_pool import toolbar_pool
 from django.core.urlresolvers import reverse
 from django.utils.translation import ugettext_lazy as _
-
 
 from .models import BLOG_CURRENT_POST_IDENTIFIER
 
@@ -22,17 +23,17 @@ class BlogToolbar(CMSToolbar):
         admin_menu.add_modal_item(_('Add post'), url=url)
 
         current_post = getattr(self.request, BLOG_CURRENT_POST_IDENTIFIER, None)
-        if current_post and self.request.user.has_perm('djangocms_blog.change_post'):   # pragma: no cover
+        if current_post and self.request.user.has_perm('djangocms_blog.change_post'):  # pragma: no cover  # NOQA
             admin_menu.add_modal_item(_('Edit Post'), reverse(
                 'admin:djangocms_blog_post_change', args=(current_post.pk,)),
                 active=True)
 
     def post_template_populate(self):
         current_post = getattr(self.request, BLOG_CURRENT_POST_IDENTIFIER, None)
-        if current_post and self.request.user.has_perm('djangocms_blog.change_post'):   # pragma: no cover
+        if current_post and self.request.user.has_perm('djangocms_blog.change_post'):  # pragma: no cover  # NOQA
             # removing page meta menu, if present, to avoid confusion
             try:   # pragma: no cover
-                import djangocms_page_meta
+                import djangocms_page_meta  # NOQA
                 menu = self.request.toolbar.get_or_create_menu('page')
                 pagemeta = menu.get_or_create_menu('pagemeta', 'meta')
                 menu.remove_item(pagemeta)
@@ -40,7 +41,7 @@ class BlogToolbar(CMSToolbar):
                 pass
             # removing page tags menu, if present, to avoid confusion
             try:   # pragma: no cover
-                import djangocms_page_tags
+                import djangocms_page_tags  # NOQA
                 menu = self.request.toolbar.get_or_create_menu('page')
                 pagetags = menu.get_or_create_menu('pagetags', 'tags')
                 menu.remove_item(pagetags)