+++ djangocms_blog/feeds.py
@@ -1,10 +1,13 @@
 # -*- coding: utf-8 -*-
+from __future__ import absolute_import, print_function, unicode_literals
+
 from django.contrib.sites.models import Site
 from django.contrib.syndication.views import Feed
 from django.core.urlresolvers import reverse
 from django.utils.translation import ugettext as _
 
 from .models import Post
+from .settings import get_setting
 
 
 class LatestEntriesFeed(Feed):
@@ -22,7 +25,9 @@ class LatestEntriesFeed(Feed):
         return item.safe_translation_getter('title')
 
     def item_description(self, item):
-        return item.safe_translation_getter('abstract')
+        if get_setting('USE_ABSTRACT'):
+            return item.safe_translation_getter('abstract')
+        return item.safe_translation_getter('post_text')
 
 
 class TagFeed(LatestEntriesFeed):