+++ djangocms_blog/forms.py
@@ -1,35 +1,18 @@
 # -*- coding: utf-8 -*-
+from __future__ import absolute_import, print_function, unicode_literals
+
 from django import forms
 from django.conf import settings
-import django_select2
-from parler.forms import TranslatableModelForm
-
-from .models import Post
+from taggit_autosuggest.widgets import TagAutoSuggest
 
 
 class LatestEntriesForm(forms.ModelForm):
-
-    class Meta:
-        widgets = {
-            'tags': django_select2.Select2MultipleWidget
-        }
+    def __init__(self, *args, **kwargs):
+        super(LatestEntriesForm, self).__init__(*args, **kwargs)
+        self.fields['tags'].widget = TagAutoSuggest('taggit.Tag')
 
     class Media:
         css = {
             'all': ('%sdjangocms_blog/css/%s' % (settings.STATIC_URL,
                                                  'djangocms_blog_admin.css'),)
         }
-
-
-class PostAdminForm(TranslatableModelForm):
-
-    def __init__(self, *args, **kwargs):
-        super(PostAdminForm, self).__init__(*args, **kwargs)
-        original_attrs = self.fields['meta_description'].widget.attrs
-        original_attrs['maxlength'] = 160
-        self.fields['meta_description'].widget = forms.TextInput(original_attrs)
-        self.fields['meta_title'].max_length = 70
-
-    class Meta:
-        model = Post
-        exclude = ()