+++ djangocms_blog/models.py
@@ -1,24 +1,24 @@
 # -*- coding: utf-8 -*-
-from cms.models import PlaceholderField, CMSPlugin
-from cmsplugin_filer_image.models import ThumbnailOption
+from __future__ import absolute_import, print_function, unicode_literals
+
+from cms.models import CMSPlugin, PlaceholderField
 from django.conf import settings as dj_settings
-from django.contrib.sites.models import Site
 from django.core.urlresolvers import reverse
 from django.db import models
 from django.utils import timezone
 from django.utils.encoding import force_text, python_2_unicode_compatible
-from django.utils.html import strip_tags, escape
+from django.utils.html import escape, strip_tags
 from django.utils.text import slugify
-from django.utils.translation import ugettext_lazy as _, get_language
+from django.utils.translation import get_language, ugettext_lazy as _
 from djangocms_text_ckeditor.fields import HTMLField
 from filer.fields.image import FilerImageField
 from meta_mixin.models import ModelMeta
-from parler.models import TranslatableModel, TranslatedFields
 from parler.managers import TranslationManager
+from parler.models import TranslatableModel, TranslatedFields
 from taggit_autosuggest.managers import TaggableManager
 
-from .settings import get_setting
 from .managers import GenericDateTaggedManager
+from .settings import get_setting
 
 BLOG_CURRENT_POST_IDENTIFIER = 'djangocms_post_current'
 
@@ -47,7 +47,16 @@ class BlogCategory(TranslatableModel):
 
     @property
     def count(self):
-        return self.blog_posts.filter(publish=True).count()
+        return self.blog_posts.published().count()
+
+    def get_absolute_url(self):
+        lang = get_language()
+        if self.has_translation(lang):
+            slug = self.safe_translation_getter('slug', language_code=lang)
+            return reverse('djangocms_blog:posts-category', kwargs={'category': slug})
+        # in case category doesn't exist in this language, gracefully fallback
+        # to posts-latest
+        return reverse('djangocms_blog:posts-latest')
 
     def __str__(self):
         return self.safe_translation_getter('name')
@@ -67,55 +76,52 @@ class Post(ModelMeta, TranslatableModel):
     Blog post
     """
     author = models.ForeignKey(dj_settings.AUTH_USER_MODEL,
-                               verbose_name=_('Author'), null=True, blank=True,
+                               verbose_name=_('author'), null=True, blank=True,
                                related_name='djangocms_blog_post_author')
 
-    date_created = models.DateTimeField(auto_now_add=True)
-    date_modified = models.DateTimeField(auto_now=True)
-    date_published = models.DateTimeField(_('Published Since'),
+    date_created = models.DateTimeField(_('created'), auto_now_add=True)
+    date_modified = models.DateTimeField(_('last modified'), auto_now=True)
+    date_published = models.DateTimeField(_('published Since'),
                                           default=timezone.now)
-    date_published_end = models.DateTimeField(_('Published Until'), null=True,
+    date_published_end = models.DateTimeField(_('published Until'), null=True,
                                               blank=True)
-    publish = models.BooleanField(_('Publish'), default=False)
-    categories = models.ManyToManyField(BlogCategory, verbose_name=_('category'),
+    publish = models.BooleanField(_('publish'), default=False)
+    categories = models.ManyToManyField('djangocms_blog.BlogCategory', verbose_name=_('category'),
                                         related_name='blog_posts',)
-    main_image = FilerImageField(verbose_name=_('Main image'), blank=True, null=True,
+    main_image = FilerImageField(verbose_name=_('main image'), blank=True, null=True,
                                  on_delete=models.SET_NULL,
                                  related_name='djangocms_blog_post_image')
-    main_image_thumbnail = models.ForeignKey(ThumbnailOption,
-                                             verbose_name=_('Main image thumbnail'),
+    main_image_thumbnail = models.ForeignKey('cmsplugin_filer_image.ThumbnailOption',
+                                             verbose_name=_('main image thumbnail'),
                                              related_name='djangocms_blog_post_thumbnail',
                                              on_delete=models.SET_NULL,
                                              blank=True, null=True)
-    main_image_full = models.ForeignKey(ThumbnailOption,
-                                        verbose_name=_('Main image full'),
+    main_image_full = models.ForeignKey('cmsplugin_filer_image.ThumbnailOption',
+                                        verbose_name=_('main image full'),
                                         related_name='djangocms_blog_post_full',
                                         on_delete=models.SET_NULL,
                                         blank=True, null=True)
-    enable_comments = models.BooleanField(
-        verbose_name=_(u'Enable comments on post'),
-        default=get_setting('ENABLE_COMMENTS')
-    )
-    sites = models.ManyToManyField(Site, verbose_name=_(u'Site(s)'), blank=True,
+    enable_comments = models.BooleanField(verbose_name=_('enable comments on post'),
+                                          default=get_setting('ENABLE_COMMENTS'))
+    sites = models.ManyToManyField('sites.Site', verbose_name=_('Site(s)'), blank=True,
                                    null=True,
-                                   help_text=_(u'Select sites in which to show the post. '
+                                   help_text=_('Select sites in which to show the post. '
                                                u'If none is set it will be '
-                                               u'visible in all the configured sites.')
-                                   )
+                                               u'visible in all the configured sites.'))
 
     translations = TranslatedFields(
-        title=models.CharField(_('Title'), max_length=255),
+        title=models.CharField(_('title'), max_length=255),
         slug=models.SlugField(_('slug'), blank=True, db_index=True),
-        abstract=HTMLField(_('Abstract')),
-        meta_description=models.TextField(verbose_name=_(u'Post meta description'),
+        abstract=HTMLField(_('abstract'), blank=True, default=''),
+        meta_description=models.TextField(verbose_name=_('post meta description'),
                                           blank=True, default=''),
-        meta_keywords=models.TextField(verbose_name=_(u'Post meta keywords'),
+        meta_keywords=models.TextField(verbose_name=_('post meta keywords'),
                                        blank=True, default=''),
-        meta_title=models.CharField(verbose_name=_(u'Post meta title'),
-                                    help_text=_(u'used in title tag and social sharing'),
+        meta_title=models.CharField(verbose_name=_('post meta title'),
+                                    help_text=_('used in title tag and social sharing'),
                                     max_length=255,
                                     blank=True, default=''),
-        post_text=HTMLField(_('Text'), default='', blank=True),
+        post_text=HTMLField(_('text'), default='', blank=True),
         meta={'unique_together': (('language_code', 'slug'),)}
     )
     content = PlaceholderField('post_content', related_name='post_content')
@@ -150,6 +156,29 @@ class Post(ModelMeta, TranslatableModel):
         'url': 'get_absolute_url',
     }
 
+    class Meta:
+        verbose_name = _('blog article')
+        verbose_name_plural = _('blog articles')
+        ordering = ('-date_published', '-date_created')
+        get_latest_by = 'date_published'
+
+    def __str__(self):
+        return self.safe_translation_getter('title')
+
+    def get_absolute_url(self):
+        kwargs = {'year': self.date_published.year,
+                  'month': '%02d' % self.date_published.month,
+                  'day': '%02d' % self.date_published.day,
+                  'slug': self.safe_translation_getter('slug',
+                                                       language_code=get_language(),
+                                                       any_language=True)}
+        return reverse('djangocms_blog:post-detail', kwargs=kwargs)
+
+    def save(self, *args, **kwargs):
+        if not self.slug and self.title:
+            self.slug = slugify(self.title)
+        super(Post, self).save(*args, **kwargs)
+
     def get_title(self):
         title = self.safe_translation_getter('meta_title', any_language=True)
         if not title:
@@ -176,34 +205,6 @@ class Post(ModelMeta, TranslatableModel):
 
     def get_author(self):
         return self.author
-
-    class Meta:
-        verbose_name = _('blog article')
-        verbose_name_plural = _('blog articles')
-        ordering = ('-date_published', '-date_created')
-        get_latest_by = 'date_published'
-
-    def __str__(self):
-        return self.safe_translation_getter('title')
-
-    def save(self, *args, **kwargs):
-        super(Post, self).save(*args, **kwargs)
-        main_lang = self.get_current_language()
-        for lang in self.get_available_languages():
-            self.set_current_language(lang)
-            if not self.slug and self.title:
-                self.slug = slugify(self.title)
-        self.set_current_language(main_lang)
-        self.save_translations()
-
-    def get_absolute_url(self):
-        kwargs = {'year': self.date_published.year,
-                  'month': '%02d' % self.date_published.month,
-                  'day': '%02d' % self.date_published.day,
-                  'slug': self.safe_translation_getter('slug',
-                                                       language_code=get_language(),
-                                                       any_language=True)}
-        return reverse('djangocms_blog:post-detail', kwargs=kwargs)
 
     def thumbnail_options(self):
         if self.main_image_thumbnail_id:
@@ -235,44 +236,50 @@ class BasePostPlugin(CMSPlugin):
         return posts
 
     def __str__(self):
-        return unicode(self.latest_posts)
+        return force_text(self.latest_posts)
 
 
 class LatestPostsPlugin(BasePostPlugin):
 
-    latest_posts = models.IntegerField(_(u'Articles'), default=get_setting('LATEST_POSTS'),
-                                       help_text=_('The number of latests articles to be displayed.'))
-    tags = models.ManyToManyField('taggit.Tag', blank=True,
-                                  help_text=_('Show only the blog articles tagged with chosen tags.'))
-    categories = models.ManyToManyField('BlogCategory', blank=True,
-                                        help_text=_('Show only the blog articles tagged with chosen categories.'))
+    latest_posts = models.IntegerField(_('articles'), default=get_setting('LATEST_POSTS'),
+                                       help_text=_('The number of latests '
+                                                   u'articles to be displayed.'))
+    tags = TaggableManager(_('filter by tag'), blank=True,
+                           help_text=_('Show only the blog articles tagged with chosen tags.'),
+                           related_name='djangocms_blog_latest_post')
+    categories = models.ManyToManyField('djangocms_blog.BlogCategory', blank=True,
+                                        verbose_name=_('filter by category'),
+                                        help_text=_('Show only the blog articles tagged '
+                                                    u'with chosen categories.'))
 
     def __str__(self):
-        return u'%s latest articles by tag' % self.latest_posts
+        return _('%s latest articles by tag') % self.latest_posts
 
     def copy_relations(self, oldinstance):
-        self.tags = oldinstance.tags.all()
+        for tag in oldinstance.tags.all():
+            self.tags.add(tag)
 
     def get_posts(self, request):
         posts = self.post_queryset(request)
-        tags = list(self.tags.all())
-        if tags:
-            posts = posts.filter(tags__in=tags)
-        return posts[:self.latest_posts]
+        if self.tags.exists():
+            posts = posts.filter(tags__in=list(self.tags.all()))
+        if self.categories.exists():
+            posts = posts.filter(categories__in=list(self.categories.all()))
+        return posts.distinct()[:self.latest_posts]
 
 
 class AuthorEntriesPlugin(BasePostPlugin):
     authors = models.ManyToManyField(
-        dj_settings.AUTH_USER_MODEL, verbose_name=_('Authors'),
+        dj_settings.AUTH_USER_MODEL, verbose_name=_('authors'),
         limit_choices_to={'djangocms_blog_post_author__publish': True}
     )
     latest_posts = models.IntegerField(
-        _(u'Articles'), default=get_setting('LATEST_POSTS'),
+        _('articles'), default=get_setting('LATEST_POSTS'),
         help_text=_('The number of author articles to be displayed.')
     )
 
     def __str__(self):
-        return u'%s latest articles by author' % self.latest_posts
+        return _('%s latest articles by author') % self.latest_posts
 
     def copy_relations(self, oldinstance):
         self.authors = oldinstance.authors.all()