+++ tests/test_plugins.py
@@ -1,10 +1,14 @@
 # -*- coding: utf-8 -*-
+from __future__ import absolute_import, print_function, unicode_literals
+
 import re
+
 from cms.api import add_plugin
 from django.core.urlresolvers import reverse
-from django.template import RequestContext
 from django.utils.timezone import now
 from taggit.models import Tag
+
+from djangocms_blog.models import BlogCategory
 
 from . import BaseTest
 
@@ -19,17 +23,35 @@ class PluginTest(BaseTest):
         post1.publish = True
         post1.save()
         ph = page1.placeholders.get(slot='content')
+
         plugin = add_plugin(ph, 'BlogLatestEntriesPlugin', language='en')
         tag = Tag.objects.get(slug='tag-1')
         plugin.tags.add(tag)
-        request = self.get_page_request(page1, self.user, r'/en/blog/', lang_code='en', edit=True)
-        context = RequestContext(request, {})
+
+        context = self.get_plugin_context(page1, 'en', plugin, edit=True)
         rendered = plugin.render_plugin(context, ph)
         self.assertTrue(rendered.find('cms_plugin-djangocms_blog-post-abstract-1') > -1)
         self.assertTrue(rendered.find(reverse('djangocms_blog:posts-tagged', kwargs={'tag': tag.slug})) > -1)
         self.assertTrue(rendered.find('<p>first line</p>') > -1)
         self.assertTrue(rendered.find('<article id="post-first-post"') > -1)
         self.assertTrue(rendered.find(post1.get_absolute_url()) > -1)
+
+        category_2 = BlogCategory.objects.create(name=u'category 2')
+        category_2.set_current_language('it', initialize=True)
+        category_2.name = u'categoria 2'
+        category_2.save()
+        category_2.set_current_language('en')
+        post2.categories.add(category_2)
+        plugin = add_plugin(ph, 'BlogLatestEntriesPlugin', language='en')
+        plugin.categories.add(category_2)
+
+        context = self.get_plugin_context(page1, 'en', plugin, edit=True)
+        rendered = plugin.render_plugin(context, ph)
+        self.assertTrue(rendered.find('cms_plugin-djangocms_blog-post-abstract-2') > -1)
+        self.assertTrue(rendered.find(reverse('djangocms_blog:posts-category', kwargs={'category': category_2.slug})) > -1)
+        self.assertTrue(rendered.find('<p>second post first line</p>') > -1)
+        self.assertTrue(rendered.find('<article id="post-second-post"') > -1)
+        self.assertTrue(rendered.find(post2.get_absolute_url()) > -1)
 
     def test_plugin_authors(self):
         page1, page2 = self.get_pages()
@@ -42,8 +64,8 @@ class PluginTest(BaseTest):
         ph = page1.placeholders.get(slot='content')
         plugin = add_plugin(ph, 'BlogAuthorPostsPlugin', language='en')
         plugin.authors.add(self.user)
-        request = self.get_page_request(page1, self.user, r'/en/blog/', lang_code='en', edit=True)
-        context = RequestContext(request, {})
+
+        context = self.get_plugin_context(page1, 'en', plugin, edit=True)
         rendered = plugin.render_plugin(context, ph)
         self.assertTrue(rendered.find(reverse('djangocms_blog:posts-author', kwargs={'username': self.user.get_username()})) > -1)
         self.assertTrue(rendered.find('2 articles') > -1)
@@ -60,9 +82,8 @@ class PluginTest(BaseTest):
         post2.save()
         ph = page1.placeholders.get(slot='content')
         plugin = add_plugin(ph, 'BlogTagsPlugin', language='en')
-        request = self.get_page_request(page1, self.user, r'/en/blog/', lang_code='en', edit=True)
-        context = RequestContext(request, {})
-        rendered = plugin.render_plugin(context, ph).replace("\n", "")
+        context = self.get_plugin_context(page1, 'en', plugin, edit=True)
+        rendered = plugin.render_plugin(context, ph)
         for tag in Tag.objects.all():
             self.assertTrue(rendered.find(reverse('djangocms_blog:posts-tagged', kwargs={'tag': tag.slug})) > -1)
             if tag.slug == 'test-tag':
@@ -81,9 +102,9 @@ class PluginTest(BaseTest):
         post2.save()
         ph = page1.placeholders.get(slot='content')
         plugin = add_plugin(ph, 'BlogCategoryPlugin', language='en')
-        request = self.get_page_request(page1, self.user, r'/en/blog/', lang_code='en', edit=True)
         plugin_class = plugin.get_plugin_class_instance()
-        context = plugin_class.render(RequestContext(request, {}), plugin, ph)
+        context = self.get_plugin_context(page1, 'en', plugin, edit=True)
+        context = plugin_class.render(context, plugin, ph)
         self.assertTrue(context['categories'])
         self.assertEqual(list(context['categories']), [self.category_1])
 
@@ -96,14 +117,15 @@ class PluginTest(BaseTest):
         post2.save()
         ph = page1.placeholders.get(slot='content')
         plugin = add_plugin(ph, 'BlogArchivePlugin', language='en')
-        request = self.get_page_request(page1, self.user, r'/en/blog/', lang_code='en', edit=True)
         plugin_class = plugin.get_plugin_class_instance()
-        context = plugin_class.render(RequestContext(request, {}), plugin, ph)
+
+        context = self.get_plugin_context(page1, 'en', plugin, edit=True)
+        context = plugin_class.render(context, plugin, ph)
         self.assertEqual(context['dates'][0]['date'].date(), now().replace(year=now().year, month=now().month, day=1).date())
         self.assertEqual(context['dates'][0]['count'], 2)
 
         post2.publish = False
         post2.save()
-        context = plugin_class.render(RequestContext(request, {}), plugin, ph)
+        context = plugin_class.render(context, plugin, ph)
         self.assertEqual(context['dates'][0]['date'].date(), now().replace(year=now().year, month=now().month, day=1).date())
         self.assertEqual(context['dates'][0]['count'], 1)