+++ tests/test_toolbar.py
@@ -1,8 +1,10 @@
 # -*- coding: utf-8 -*-
+from __future__ import absolute_import, print_function, unicode_literals
+
 from cms.toolbar.items import ModalItem
 from django.core.urlresolvers import reverse
-from djangocms_blog.models import BLOG_CURRENT_POST_IDENTIFIER
 
+from djangocms_blog.models import BLOG_CURRENT_POST_IDENTIFIER
 
 from . import BaseTest
 
@@ -24,4 +26,3 @@ class ToolbarTest(BaseTest):
         self.assertEqual(len(blog_menu.find_items(ModalItem, url=reverse('admin:djangocms_blog_post_changelist'))), 1)
         self.assertEqual(len(blog_menu.find_items(ModalItem, url=reverse('admin:djangocms_blog_post_add'))), 1)
         self.assertEqual(len(blog_menu.find_items(ModalItem, url=reverse('admin:djangocms_blog_post_change', args=(post.pk,)))), 1)
-